+++ packages/core/babel-preset-dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # babel-preset-dlight
 
+## 1.0.0-beta.6
+
+### Patch Changes
+
+- feat: add @Mount auto render entry
+
 ## 1.0.0-beta.5
 
 ### Patch Changes