+++ packages/core/dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/dlight
 
+## 1.0.0-beta.9
+
+### Patch Changes
+
+- feat: add @Mount auto render entry
+
 ## 1.0.0-beta.8
 
 ### Patch Changes